requirements.txt
python-dotenv==1.0.1
requests==2.32.3
ytmusicapi==1.8.0